Add HorizontalPageIndicator
HorizontalPageIndicator represents a total number of pages and a selected page. Recommended way to use this control is to use it with an Accompanist HorizontalPager (or any pager in general), however currently there's no compose Pager in AndroidX, that's why a demo with Pager was not included into demos. It will be included into Compose-advanced and Horologist projects

Bug: 218669425
Test: wear.compose.material.HorizontalPageIndicatorTest.kt
Relnote: "Add HorizontalPageIndicator"
Change-Id: Iac8984eb938e64816850c494ffcaa7988e724cc8