Convert glance template layouts to use a mapping function
Rather than a set of abstract functions for each layout type, templates
use a map from display context to layout. Currently the display context
only includes display orientation. Widgets can use the base function and
get the correct mapping for their context.

This mapping is preferred so that surface-specific sub-libraries can create and
inject these layouts. This means the layouts are not created in the
top-level library, so they can use surface-specific features without the
need to depend on all the glance surface libraries.

Relnote: Convert glance template layouts to use a map
Test: Demo added
Change-Id: I46bfde93ec59407940b40f8e8fb2790b7509ac11