Update PickerState such that NumberOfOptions may be updated
PickerState now takes the initial number of options in the constructor and the numberOfOptions is settable - this supports use-cases such as a DatePicker, when the number of days in the month changes depending on the month selected. Also updated the DatePicker demo to take account of the maximum number of days in each month.

Test: androidx.wear.compose.material tests and androidx.wear.compose.integration.demo tests

Relnote: "We have updated PickerState so that the numberOfOptions can be updated. This supports use-cases such as a DatePicker, when the number of days in the month changes depending on the month selected. The constructor parameter for PickerState has changed to initialNumberOfOptions accordingly."

Bug: 220859488
Change-Id: Iad066cbe48265179e09cf17401e4a42344e09f5d