Replace usages of Kotlin's `use` with our own `useCursor` that is backwards compatible.
Its not until API 16 that android.database.Cursor implements Closeable so using `kotlin.io.use()` in lower APIs will cause a ClassCastException. This change replaces usages of Kotlin's `use()` with `useCursor()` that is backwards compatible.

Additionally, to prevent future regressions, this change adds a custom lint rule preventing the usage of `kotlin.io.use()` with `Cursor` as its receiver type. The room-runtime-lint artifact is unpublished and the checks are local to the androidx codebase.

Bug: 223667890
Test: :room:integration-tests:room-testapp:cAT on API 15
Change-Id: I6c3b1d63593ddb257ddbd470cc9b1a739f31ead1