Update Picker and Typography demos and samples
See bugs for details, main Picker is to remove OptionChangePicker and DualPicker samples from the Demo app and add a test rig so that those samples are referenced and continue to build. For typography, added a sample for how to fix the font size so that they are not scaled according to the system font size settings (which can be necessary for large font sizes which are already sufficiently large for accessibility).

Test: androidx.wear.compose.integration.demos.test and androidx.wear.compose.material.samples.test
Bug: 223808388
Bug: 223819823

Relnote: N/A
Change-Id: I6afecc0eee18ed7dd99ba2b2ff429c6f54556ae2