Wrap CheckBox in a FrameLayout
This change updates the checkbox template so that it is wrapped in a
FrameLayout. This allows us to achieve our desired padding behavior,
where the padding is applied to the text and button as a unit. When
padding is set on the CheckBox view alone, it is only applied to the
text (see bug).

Bug: 215307943
Test: CheckBoxTranslatorTest, integration tests
Change-Id: I31a19d5722c6e6544285a12ea2aed1d8abc0a225