Fix corner a case where the composer would crash
If a group with nodes is deleted in front of a group that is doesn't contain
nodes and is moved to new location and nodes are added to it could cause the
composer to crash with an invalid start index.

The most likely case that can cause this is a for loop with nested if
statements like the included test case.

Test: ./gradlew :comopse:r:r:tDUT
Fixes: 199136503
Change-Id: I89f6abce423030d4345351d2533f867fd1cb519f