Add Decimal as a new KeyboardType
Most keyboards show a decimal separator when keyboard type is set as Number. However, it is possible that a keyboard expects `TYPE_NUMBER_FLAG_DECIMAL` flag in inputType to actually show a key for decimal separator.

This change adds a new KeyboardType called Decimal that explicitly sets the required flag. Number and Decimal would essentially behave the same for most keyboards and OEMs.

Fix: b/209835363
Test: EditorInfoTest
Relnote: "Added KeyboardType.Decimal as an alternative to Keyboard.Number for specifically including decimal separator in IME."
Change-Id: Iec4c84be81e72f8eaf136f2df23f9d12567cc3dc