Add overload with onDismissed to SwipeToDismissBox
The most common usage of SwipeToDismissBox is to define a LaunchedEffect which snaps the offset back to the default position and makes a call to navigate. Here, we are adding an overload to support that usage directly, rather than providing it in our SimpleSwipeToDismissBox sample.

Test: androidx.wear.compose.integration.demos.test and androidx.wear.compose.material.samples.test
Bug: 226565726

Relnote: "We have added a SwipeToDismissBox overload with an onDismissed parameter to support the common usage of triggering a navigation event when the swipe gesture has completed."
Change-Id: I503534f8d11045c9446e976764134c05e147ea33