Removes the cppFlags from local-storage build.gradle.
CppFlags had to be removed from android.defaultConfig.externalNativeBuild.cmake.
This is because in CMake 3.22.1, this overwrites all cpp flags when building libicing.
By overwriting these flags, the binary jumped from 3.7MB to 9.7MB.
This will bring the binary size back down to 3.7MB.
The C version will be set in CMakeLists.txt.

Test: Run AppSearch-Local-Storage tests
Change-Id: I3636f4919814a0cdf2a3d624bd682845791f9fe0