Fix ArrayMap.toArray
This CL fixes a bug in ArrayMap's toArray methods where it wasn't
properly copying from its super class. It happened because it used
the package private backing array from the super, which is not accessible
anymore after the kotlin migration.

Instead of making it internal, I just removed the helper method and
inlined it in the toArray implementations. The method itself was
confusing to read w/ the offset anyways.

Bug: 228109589
Test: ArrayMapTest.toArray
Change-Id: Ib348bfd61eb4e6a152192755baabded26804de22