Add padding modifier for curved elements.
Implement padding modifiers. Currently only absolute paddings can be
specified, but the API is there to introduce layout direction (curved or
otherwise) aware paddings.
As part of this, created a new CurvedMeasureScope, that is a density but
also provides access to the curved layout direction to CurvedChild and
CurvedModifiers.

RelNote: "Added CurvedModifier.padding* functions"
Test: ./gradlew :wear:compose:compose-material:connectedCheck --info --daemon
Bug: 227444035

Change-Id: I4dbb48de80390332def76f830b1802167bd8df15