WearCompose Pre-Beta API Clean up
Add ColumnScope/RowScope/BoxScope to slots where it makes sense to indicate to developers what the layout assumptions for the slots are. This will allow developer to use additional modifiers and avoid providing extra layout elements in some situations.

Changed the ordering and defaulting of Card colors so that most of the colors follow contentColor so the developer only needs to override this 1 value if they want to customize the card colors.

Change fun ToggleChipDefaults.switchUncheckedIconColor() to val ToggleChipDefaults.SwitchUncheckedIconColor


Test: ./gradlew :wear:compose:compose-material:connectedCheck --info --daemon
RelNote: "We have added RowScope/ColumnScope/BoxScope to some of the slots in our Composables to indicator to developers what the layouts assumptions are. This will allow developers to use additional modifiers on some slot content and avoid the need to provide extra layout elements. Additionally we have made some minor updates to the AppCard/TitleCard colors so that timeColor and appColor default to contentColor, these properties can still all be individually overridden if needed."
Bug: 228940380
Change-Id: I26b597bb6b2c9db3d8d3833a2dfa36374c7631a2