Wear Compose rename toggleControlTintColor/iconTintColor to toggleControlColor/iconColor
iconTintColor and toggleControlTintColor are used to color the slots used for icons and toggleControls. The use of tint in the name isn't helpful and is potentially misleading as actually the color specified will be applied to whatever content goes into the slot.

Test: ./gradlew :wear:compose:compose-material:connectedCheck --info --daemon
RelNote: "We have renamed iconTintColor and toggleControlTintColor to iconColor and toggleControlColor throughout the API (Chip/ToggleChip/Dialog/Slider/Stepper/...) as the color is applied to the icon/toggleControl slot."
Bug: 229058667
Change-Id: Ied23864a671b3cfd241cb0f077f111b762525a33