Refactor Room compiler cursor index variable generation to its own adapter.
This changes introduces IndexAdapter that is able to create local variables containing the indices of the columns to read from a cursor. This logic used to be part of the RowAdapter but as part of multimap index resolution, it seems best to extract it out such that more sophisticated index adapters are used that take into consideration multiple row adapters along with their mappings.

The coupling between RowAdapter and IndexAdapter is still important and as follow:
* A RowAdapter must provide an IndexAdapter implementation by default
* All IndexAdapters' onCursorReady() must be invoked before RowAdapters' onCursorReady()
* The IndexAdapter generated local variable names containing the column indices must be passed to the RowAdapter for it to properly convert a cursor to an object.

Test: Existing
Change-Id: If1792af101fa80db6749de711aba4871c44b1746