Rework how we saved state in FragmentStore
Now that we have moved the list of FragmentStates
out of the FragmentManagerState and into the
Fragment bundle, we have direct access to each
fragment's saved state and should utilize that
in FragmentStore. Instead of keep track of a list
of FragmentStates, we can simple keep a map of the
fragment's mWho to its Fragment Bundle and use that
to access its saved state. This will become particularly
useful as we move the user states from FragmentState to
the Fragment bundle.

Test: ran all tests in androidx.fragment.app
Bug: 207158202
Change-Id: I800ebe9d9ec9cf9a8c0179decc735ea202e55da7