Use signature protectionLevel for unexported runtime receivers
Android T introduced support for runtime receivers registering for
non-system broadcasts to be registered as unexported to other apps
on the device via a new RECEIVER_NOT_EXPORTED flag. Support was added
to ContextCompat to allow developers to use similar flags when
registering their receivers; on T+ devices, the flag would be
forwarded to the platform, and on pre-T devices a receiver registered
as unexported would be protected with a permission. This commit
updates the permission declaration for pre-T devices with the
signature protectionLevel to ensure only apps signed with the
registering app's signing key can send broadcasts to these receivers.

Bug: 199978728
Test: ContextCompatTest
Test: gradlew createArchive and used the built repo from an app
      to ensure the receiver was protected with a signature permission
Change-Id: I55544c62675102db80a9c17310a5605dd893608e