Stop using CompositionData.asTree()
Before this change we used: `CompositionData.asTree()` API which
makes a complete copy of the SlotTree into temporary tree of Group
nodes. Each node in the tree has an encoded `sourceInformation`
string and all such strings will be fully parsed using this method.
We then took the Group tree and created a smaller tree of the nodes
we want to display in the Layout Inspector.

Now we use: `CompositionData.asLazyTree()` instead, which allows:
- caching parsed `sourceInformation` (these are not unique)
- avoiding the computation of source location if not needed
- avoiding the extraction of parameters if not needed
- avoiding parsing when only the name of the composable is needed

And add options for (but do not activate):
- delaying parameter extraction
- filtering out nodes outside of the android View

The logic for sub compositions unfortunately also needed to be
changed slightly because a CompositionGroup returns different child
nodes every time `CompositionGroup.compositionGroups` is accessed.

Bug: 229657989
Test: Existing tests must still pass.
Change-Id: Ibbd985e869914df0aeb41563fcc6a505c511449b