Fix fail to remove VisibilitySettings when remove a schema.
When we remove the schema from the AppSearch, we also need to remove
the visibility setting of that schema. But if the visibility setting
is not all default. we will fail to remove the visibility setting,
and if the user adds that schema back with an all default setting,
the old setting will be picked up.

Bug: 230280377
Test: AppSearchImplTest, VisibilityStoreTest
Change-Id: I807b180c6d721af22212c10a9842371257e98d90