Revert^3 "Adding implementation for Uwb ranging"
d6f5561c1bee2d704a39a65b07da33876c8d72ca

Revert reason: checked in prebuilts are overriding a version that
is a version that already exists and it does not match the existing
maven artifact content of `com.google.android.gms:play-services-nearby:18.1.0`.
It is not clear if this is a locally built or a build server version.

Change-Id: Ic6691da3bf7a09e2af8e7db5982e14f9d11f7b37