Picker should pad content when gradient is present.
There is jittering (artefacts left on the screen) during scaling when a gradient is shown on the Picker (this happens during SwipeToDismiss). It can be prevented by padding within the gradient region by 1.dp.

Bug: 223386180
Test: Manual testing in androidx.wear.compose.integration.demos
Relnote: "We have added 1dp padding around Picker contents when drawn with a gradient to prevent jitter on text seen when swiping."
Change-Id: I0b7b946c55a3553bd869a2d2add2958fd909a591