Stop automatically rewinding in `scheduleInternal` in `TestScheduler`
That behavior resulted in immediate attempts to re-run workers that finished
with `Result.retry()`. Initially `rewindLastEnqueueTime` was moved to
`scheduleInternal` to fix the bug, when PeriodicWorkers with flex weren't run
immediately after enqueue in TestScheduler. Now we rewind time in schedule
only if workRequest has `periodDelayMet` equal to true. This only happens on the
first period.

bug: 230503437
Test: TestSchedulerTest
Change-Id: If72f258aff20ca24069cef126e993811a6cbe46c