Restructure WebViewFeatureInternal to facilitate static analysis
In order to allow for static analysis of NewApi usage in code protected
by the isSupportedByFramework method, a new value class (ApiFeature) is
introduced, which is subclassed for each SDK version.
WebViewFeatureInternal is changed from an enum to a regular class with
constants of type ApiFeature.X where X is the SDK level that supports
the individual feature.

BYPASS_INCLUSIVE_LANGUAGE_REASON=Refactoring existing public API

Bug: 204918082
Test: ./gradlew :webkit:webkit:lint
Test: ./gradlew :webkit:webkit:connectedAndroid
Change-Id: I2f2d4d9a8bfb4f699cd0d00b07d410189353e6fc