Picker should respond to scrolling even in read-only mode.
Picker has a read-only mode that shows only the currently selected option. For multi-picker screens, it is necessary to tap the Picker to select it and then scroll. We want to update it so that the Picker always responds to scrolling. With this change, the Picker remains scrollable in read-only mode and options other than that currently selected are obscured by a shim in gradientColor.

Bug: 223386176
Test: androidx.wear.compose.material
Relnote: "Picker now always responds to scroll events even when in read-only mode. This avoids the need for users to first select a Picker by tapping, before they can scroll. In read-only mode, options other than that currently selected are obscured by a shim in gradientColor."
Change-Id: I72925a66999fa5ad1ec2307189483484aae4c9fc