Allow for concurrent modification of delegators list
JankStats allows multiple JankStats instances to be active on the same
window, each with their own listener receiving metrics which are delegated
from a single listener inside of JankStats. But the thread issuing metrics
(the UI thread pre-24, or the FrameMetrics thread in 24+) may not be the
same thread in which JankStats are created, which add to the list of
delegates, causing potential concurrent modifications as delegates are
added or removed from the same list that is being iterated through
on a different thread to send metrics.

The fix is to synchronize around a common object which the delegates use.
For pre-24, we use the DecorView, which is where the tag is set to stash
the root listener which delegates to the list of delegates. On 24+ we
use the Window owning that DecorView (because it is more accessible in the
places it is needed). Note that we only need to synchronize at the root of
all code paths accessing the delegates list, which is during processing of
a frame callback (OnPreDrawListener pre-24 or FrameMetricsListener 24+),
as well as setupFrameTimer(), which the calls methods to add or remove
delegates.

Along the way, fixed a bug in JankStatsApi16Impl which prevented delegates
from being removed at all (now also tested in
JankStatsTest#testMultipleListeners())

Bug: 230388846
Test: Added to JankStatsTest#testMultipleListeners, ran on API 17, 24, and 32

Change-Id: I2827f94eeee646e8d2c439c16997a2950fe80a10