Fix picker/list snap
Previous implentation used a spring for the final snap, that with the
used parameters toke a long time to finish (evemo more than 2 seconds),
and while the fling is in progress, the UI doesn't accept taps.
Also, we now check if the scroll is actually consumed, to avoid trying
to scroll beyond the end of the content.
Test: ./gradlew :wear:compose:compose-material:connectedCheck --info --daemon
Bug: 223384491

Change-Id: I73db72901c798e371ab091a276515299ded7059e