Enable commonization for KMP projects
This flag is necessary for collections kmp which relies on it
for its posix calls.
Previously, it was only added to the github build because we
run KMP only there but we also need it in aosp as we want to
run KMP collections in AOSP once we can figure out the build
machine story.

Bug: 232103631
Test: development/update-verification-metadata.sh
Change-Id: Ia6563fb812560bb2d5857718dbc2640b11298b76