Add a demo for animating the start text of a TimeText
The demo showcases a way to do the animated fade in/grow and fade out
shrink of the text before the time in a TimeText, similar to UX specs.
Also, improve the code for ellipsising/cliping, as this demo showed
it was not great (it could be greatly simplified for the ellipsising,
since code already existed in android's TimeUtil)
Minor change to improve the performance avoiding the recreation of the
CurvedTextDelegate.
Relnote: "New demo to animate adding or removing a start text on a
TimeText."
Test: ./gradlew :wear:compose:compose-material:connectedCheck --info --daemon
Bug: 232783349

Change-Id: I16d75a1b694b50fb625b75a3301f335318d4490e