Add keyboardModifiers field to WindowInfo
This will allow reading the state of keyboardModifiers using LocalWindowInfo.
It is helpful in cases when users need to know the state of keyboardModifiers, for example: in the callbacks of drag listeners, or when pointerIcon depends on the state of keyboardModifiers.
This is especially important for the drag gesture, when position doesn't change (thus onDrag is not called) but keyboardModifiers change.

Relnote: Added a new API WindowInfo.keyboardModifiers to observe its state within composable functions or via snapshotFlow

Test: added a test windowInfo_providesKeyModifiers in WindowInfoCompositionLocalTest.kt
Change-Id: Icdb8a0e53f5adb436662bc62c620543e77029f1a