Fix namespacing for Navigation SafeArgs in AGP
There is an issue with the new namespace support in Navigation Safe Args
which if there is an applicationIdSuffix and a namespace with no package
name provided in an AndroidManifest, or there is a suffix with differing
namespace and package name in AndroidManifest, the resource directory
package isn't properly generated in the SafeArgs class.

Namespace does not take into account applicationIdSuffix and SafeArgs
still needs to since the generated classes need to be in the same package
as the classes they are being generated from if they use relatives paths
in the xml.

This means the following:
1. We always need to use the applicationId to get the proper location of
   the generated file.
2. If a namespace is available, we need to use that to identify the
   location of the R resource class.

RelNote: "`Navigation SafeArgs` will no longer crash when using an
`applicationIdSuffix` and namespace with either no applicationId
present, or when the applicationId and namespace differ."
Test: modified safeargs test
Bug: 233119646

Change-Id: I754b1fffe44bdb142ae818187cad6cd7fdc9080a