Workaround ViewModel crash with compile only dependencies
AGP currently does not properly handle desugaring compileOnly
dependencies. This causes issues with the new version of the create()
function in ViewModelProvider that results in an AbstractMethodError.
This breaks devices using any API under 24 that attempt to use new
libraries that depend on the current version of lifecycle.

We can workaround this so that we can at least keep devs from crashing
if their compileOnly dependency is older.

RelNote: "`ViewModelProvider` will no longer crash when mixing previous
versions of compileOnly Lifecycle dependencies with versions 2.5+."
Bug: 230454566
Test: ./gradlew bOS

Change-Id: I81a66ddfc4bc1fb9a98fa8fa526176a8506e1ed8