Implement signal generator and video recorder
The test app is served as both a signal generator and a video
recorder. The signal generator produces periodic light and sound
signals. The recorder record videos with CameraX's VideoCapture.

The light source is achieved by setting the screen background color
to white to illuminate the environment. The sound source is simulated
by using a single frequency sine wave.

The app does not handle the permission settings, because it is expected
to be run from command line by E2E test, in which case the permissions
are already being handled.

The frequency of the beep sound can be controlled by setting value to
the key `beep_frequency` while starting the app.

See go/CameraX-av-sync-test for more information.

Bug: 234100453
Test: manual test and ./gradlew bOS
Change-Id: Ia2aa2440265e5343fc360a62b89e51d425396e93