Fix ScalingLazyColumn initialization bug if the 0th list items was large
If the 0th list item in a ScalingLazyColumn is large enough it is possible that the list will not draw correctly until after the list has been scrolled. This is caused by the ScalingLazyColumn initially drawing items transparently in order to then do a correct scroll to the required initial item index without the user seeing the scroll. The calculation to determine whether to draw the items transparently or not was failing to correctly take into account the gap between list items and so could incorrectly determine that the list was still uninitialized.

Bug: 234328517
Test: ./gradlew :wear:compose:compose-material:connectedCheck --info --daemon
RelNote: "We have fixed a bug in the ScalingLazyColumn that could result in the list items not drawing correctly until scrolled if the 0th list item was large enough (including padding)"
Change-Id: Ic6159b18937d97f9bb580cf59810880288ad6d4e