Add temporary access to XProcessingEnv from XConverters.
This is useful for Dagger, while we try to add missing features in
XProcessing. In particular, the XProcessingEnv is needed in static
contexts to get the current backend (javac vs ksp) and to switch back
and forth between XProcessing types. These methods are marked as
deprecated as we expect them to be removed once the missing features
are supported by XProcessing.

Test: ./room/gradlew -p room :room:r-c-p:test
Change-Id: I0a951b4b9f96b9b3baf7f98b1765e134d290b51e